Better IO IDs

+++ src/main/java/be/webtechie/vaadin/pi4j/config/CrowPi1Config.java
@@ -23,7 +23,7 @@ public class CrowPi1Config implements CrowPiConfig {
     }
 
     @Override
-    public int getI2cDeviceLcd() {
+    public byte getI2cDeviceLcd() {
         return 0x27;
     }
